--- a/src/ast.rs
+++ b/src/ast.rs
@@ -207,6 +207,7 @@ pub enum Expression {
     Bool(Bool),
     For(Box<For>),
     ForN(Box<ForN>),
+    Sum(Box<ForN>),
     If(Box<If>),
     Compare(Box<Compare>),
     UnOp(Box<UnOpExpression>),
@@ -307,9 +308,13 @@ impl Expression {
                 convert.update(range);
                 result = Some(Expression::For(Box::new(val)));
             } else if let Ok((range, val)) = ForN::from_meta_data(
-                    convert, ignored) {
+                    "for_n", convert, ignored) {
                 convert.update(range);
                 result = Some(Expression::ForN(Box::new(val)));
+            } else if let Ok((range, val)) = ForN::from_meta_data(
+                    "sum", convert, ignored) {
+                convert.update(range);
+                result = Some(Expression::Sum(Box::new(val)));
             } else if let Ok((range, val)) = Loop::from_meta_data(
                     convert, ignored) {
                 convert.update(range);
@@ -361,6 +366,7 @@ impl Expression {
             Bool(ref b) => b.source_range,
             For(ref for_expr) => for_expr.source_range,
             ForN(ref for_n_expr) => for_n_expr.source_range,
+            Sum(ref for_n_expr) => for_n_expr.source_range,
             If(ref if_expr) => if_expr.source_range,
             Compare(ref comp) => comp.source_range,
             UnOp(ref unop) => unop.source_range,
@@ -1256,11 +1262,11 @@ pub struct ForN {
 
 impl ForN {
     pub fn from_meta_data(
+        node: &str,
         mut convert: Convert,
         ignored: &mut Vec<Range>)
     -> Result<(Range, ForN), ()> {
         let start = convert.clone();
-        let node = "for_n";
         let start_range = try!(convert.start_node(node));
         convert.update(start_range);