@dobkeratops 2016-05-17
No idea how this language engine works, but do you think you could go as far as inferring the range from the mere mention of the variables. e.g. "sum(i)  {foo[i]*bar[i]} " infers the range of [i] from its use indexing foo, bar. (whatever syntax makes sense).  This would take it beyond whats possible with high-order functions and macros in existing languages; and you could still let the user over-ride some specific range sum(i=0..n-1), whatever.  This maybe way beyond what you need of course, and way too much effort